python/hand-keypoints/requirements.txt
mediapipe==0.10.32
numpy==2.4.2
opencv_contrib_python==4.11.0.86
opencv_python==4.9.0.80
websocket_client==1.7.0
